Updated to 3.5.2

+++ facebook/src/com/facebook/internal/FileLruCache.java
@@ -78,10 +78,10 @@ public final class FileLruCache {
         this.lock = new Object();
 
         // Ensure the cache dir exists
-        this.directory.mkdirs();
-
-        // Remove any stale partially-written files from a previous run
-        BufferFile.deleteAll(this.directory);
+        if (this.directory.mkdirs() || this.directory.isDirectory()) {
+            // Remove any stale partially-written files from a previous run
+            BufferFile.deleteAll(this.directory);
+        }
     }
 
     // This is not robust to files changing dynamically underneath it and should therefore only be used
@@ -102,8 +102,10 @@ public final class FileLruCache {
 
         File[] files = this.directory.listFiles();
         long total = 0;
-        for (File file : files) {
-            total += file.length();
+        if (files != null) {
+            for (File file : files) {
+                total += file.length();
+            }
         }
         return total;
     }
@@ -221,14 +223,16 @@ public final class FileLruCache {
         // get the current directory listing of files to delete
         final File[] filesToDelete = directory.listFiles(BufferFile.excludeBufferFiles());
         lastClearCacheTime.set(System.currentTimeMillis());
-        Settings.getExecutor().execute(new Runnable() {
-            @Override
-            public void run() {
-                for (File file : filesToDelete) {
-                    file.delete();
+        if (filesToDelete != null) {
+            Settings.getExecutor().execute(new Runnable() {
+                @Override
+                public void run() {
+                    for (File file : filesToDelete) {
+                        file.delete();
+                    }
                 }
-            }
-        });
+            });
+        }
     }
 
     private void renameToTargetAndTrim(String key, File buffer) {
@@ -278,14 +282,17 @@ public final class FileLruCache {
             PriorityQueue<ModifiedFile> heap = new PriorityQueue<ModifiedFile>();
             long size = 0;
             long count = 0;
-            for (File file : this.directory.listFiles(BufferFile.excludeBufferFiles())) {
-                ModifiedFile modified = new ModifiedFile(file);
-                heap.add(modified);
-                Logger.log(LoggingBehavior.CACHE, TAG, "  trim considering time=" + Long.valueOf(modified.getModified())
-                        + " name=" + modified.getFile().getName());
+            File[] filesToTrim =this.directory.listFiles(BufferFile.excludeBufferFiles());
+            if (filesToTrim != null) {
+                for (File file : filesToTrim) {
+                    ModifiedFile modified = new ModifiedFile(file);
+                    heap.add(modified);
+                    Logger.log(LoggingBehavior.CACHE, TAG, "  trim considering time=" + Long.valueOf(modified.getModified())
+                            + " name=" + modified.getFile().getName());
 
-                size += file.length();
-                count++;
+                    size += file.length();
+                    count++;
+                }
             }
 
             while ((size > limits.getByteCount()) || (count > limits.getFileCount())) {
@@ -319,8 +326,11 @@ public final class FileLruCache {
         };
 
         static void deleteAll(final File root) {
-            for (File file : root.listFiles(excludeNonBufferFiles())) {
-                file.delete();
+            File[] filesToDelete = root.listFiles(excludeNonBufferFiles());
+            if (filesToDelete != null) {
+                for (File file : filesToDelete) {
+                    file.delete();
+                }
             }
         }
